Option to export textkeys into a file
Add an option when exporting the .dae file, it also exports the textkeys as needed by OpenMW. It's a .txt file with lines of what animation starts / ends / loops happens along with other events like sounds etc. Textkeys are created from timeline markers which need to be named in a way to allow the output to work for textkeys purposes. Frame numbers are added automatically.