Material export now based on Specular BSDF
Until now the exporter took the "Principles BSDF" shader node of the active material and used that to write COLLADA material values. It was only half working, as the principled shader is meant for PBR and its values don't match what COLLADA supports. Due to that, quite a few of the exported values were hard-coded and not exposed to the user.

This change makes it so that the "Specular BSDF" shader node and its values are used instead. The textures also work (though for OpenMW only the diffuse slot needs to be taken). Haven't done extensive testing yet.